Reduce warnings on windows.
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@28216 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/avt/DBAtts/SIL/avtSILArray.C
@@ -457,7 +457,7 @@ avtSILArray::GetSetIndex(const std::string &name) const
         {
             if (names[ii] == name)
             {
-                return (iFirstSet + ii);
+                return (iFirstSet + (int)ii);
             }
         }
         return -1;
